+++ external/vulkancts/modules/vulkan/draw/vktDrawImageObjectUtil.cpp
@@ -387,7 +387,7 @@ void Image::readUsingBuffer (vk::VkQueue				queue,
 	}
 
 	deUint8* destPtr = reinterpret_cast<deUint8*>(stagingResource->getBoundMemory().getHostPtr());
-	deMemcpy(data, destPtr, bufferSize);
+	deMemcpy(data, destPtr, static_cast<size_t>(bufferSize));
 }
 
 tcu::ConstPixelBufferAccess Image::readSurfaceLinear (vk::VkOffset3D				offset,
@@ -714,7 +714,7 @@ void Image::uploadUsingBuffer (vk::VkQueue					queue,
 	BufferCreateInfo stagingBufferResourceCreateInfo(bufferSize, vk::VK_BUFFER_USAGE_TRANSFER_DST_BIT | vk::VK_BUFFER_USAGE_TRANSFER_SRC_BIT);
 	stagingResource = Buffer::createAndAlloc(m_vk, m_device, stagingBufferResourceCreateInfo, allocator, vk::MemoryRequirement::HostVisible);
 	deUint8* destPtr = reinterpret_cast<deUint8*>(stagingResource->getBoundMemory().getHostPtr());
-	deMemcpy(destPtr, data, bufferSize);
+	deMemcpy(destPtr, data, static_cast<size_t>(bufferSize));
 	vk::flushMappedMemoryRange(m_vk, m_device, stagingResource->getBoundMemory().getMemory(), stagingResource->getBoundMemory().getOffset(), bufferSize);
 	{
 		//todo [scygan] get proper queueFamilyIndex
