+++ external/vulkancts/modules/vulkan/dynamic_state/vktDynamicStateRSTests.cpp
@@ -237,7 +237,7 @@ protected:
 			m_context.getDefaultAllocator(), vk::MemoryRequirement::HostVisible);
 
 		deUint8* ptr = reinterpret_cast<unsigned char *>(m_vertexBuffer->getBoundMemory().getHostPtr());
-		deMemcpy(ptr, &m_data[0], dataSize);
+		deMemcpy(ptr, &m_data[0], static_cast<size_t>(dataSize));
 
 		vk::flushMappedMemoryRange(m_vk, device,
 								   m_vertexBuffer->getBoundMemory().getMemory(),
